server/requirements.txt
flask
requests
python-dotenv
flask-cors
APScheduler
sqlalchemy
pydantic
feedparser
openai
psycopg2-binary
bcrypt
beautifulsoup4
lxml
pandas
xlrd
openpyxl